requirements.txt
# NK Immunosenescence Pipeline - Python Dependencies
# Pinned versions for reproducibility
# Last updated: 2025-01-03

# Core scientific computing
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
matplotlib==3.9.1
seaborn==0.13.2

# Single-cell analysis ecosystem
scanpy==1.10.2
anndata==0.10.8
scvi-tools==1.1.5
scib==1.1.5
scib-metrics==0.5.1
mudata==0.2.4
pegasuspy==1.10.0

# Machine learning and deep learning
scikit-learn==1.5.1
torch==2.4.0
pytorch-lightning==2.3.3
jax[cuda12]==0.4.30
lightgbm==4.4.0

# GPU acceleration
faiss-gpu==1.7.2

# Network analysis and clustering
python-igraph==0.11.6
leidenalg==0.10.2
louvain==0.8.2

# Statistical analysis
statsmodels==0.14.2
patsy==0.5.6
scikit-misc==0.3.1

# Doublet detection
scrublet==0.2.3

# Gene set enrichment
gseapy==1.1.3

# Data visualization
plotly==5.22.0
umap-learn==0.5.6

# Sparse matrix handling
sparse==0.15.4

# Jupyter and interactive computing
jupyter==1.0.0
jupyterlab==4.2.3
ipywidgets==8.1.3
ipykernel==6.29.5

# File I/O and data formats
h5py==3.11.0
tables==3.9.2
openpyxl==3.1.5

# Progress bars
tqdm==4.66.4

# YAML configuration
pyyaml==6.0.1

# Logging and utilities
colorlog==6.8.2

# Documentation
sphinx==7.4.6
sphinx-rtd-theme==2.0.0
sphinx-autodoc-typehints==2.2.2

# Testing
pytest==8.2.2
pytest-cov==5.0.0

# Code quality
black==24.4.2
flake8==7.1.0
mypy==1.10.1

# CELLxGENE Census data access
cellxgene-census==1.14.1

# R interface for DESeq2 (pseudobulk DE)
rpy2==3.5.16

# Additional utilities
joblib==1.4.2
numba==0.60.0
memory-profiler==0.61.0

# CUDA support (already in base image but explicit for clarity)
# nvidia-cudnn-cu12==8.9.7.29
# nvidia-cuda-runtime-cu12==12.1.105
